+++ waves/tests/test_custom_study.py
@@ -69,7 +69,6 @@ class TestCustomStudy:
                              ids=generate_io.keys())
     def test_generate(self, parameter_schema, expected_array):
         TestGenerate = CustomStudy(parameter_schema)
-        TestGenerate.generate()
         generate_array = TestGenerate._samples
         assert numpy.all(generate_array == expected_array)
         # Verify that the parameter set name creation method was called
@@ -107,10 +106,8 @@ class TestCustomStudy:
                              ids=merge_test.keys())
     def test_merge(self, first_schema, second_schema, expected_array):
         TestMerge1 = CustomStudy(first_schema)
-        TestMerge1.generate()
         with patch('xarray.open_dataset', return_value=TestMerge1.parameter_study):
             TestMerge2 = CustomStudy(second_schema, previous_parameter_study='dummy_string')
-            TestMerge2.generate()
         generate_array = TestMerge2._samples
         assert numpy.all(generate_array == expected_array)
         # Check for consistent hash-parameter set relationships
@@ -169,7 +166,6 @@ class TestCustomStudy:
                                         output_file_template=output_file_template,
                                         output_file=output_file,
                                         output_file_type=output_type)
-            TestWriteYAML.generate()
             TestWriteYAML.write()
             stdout_write.assert_not_called()
             xarray_to_netcdf.assert_not_called()